LS: Move diagnostics refreshing to separate module
commit-id:4b7d9432
