+++ crates/cairo-lang-language-server/src/lang/diagnostics/mod.rs
@@ -1,1 +1,3 @@
 pub mod lsp;
+// TODO(mkaput): This is pub temporarily.
+pub mod refresh;
