+++ crates/cairo-lang-language-server/src/lib.rs
@@ -38,9 +38,8 @@
 //! }
 //! ```
 
-use std::collections::{HashMap, HashSet};
 use std::io;
-use std::panic::{AssertUnwindSafe, RefUnwindSafe, catch_unwind};
+use std::panic::RefUnwindSafe;
 use std::path::{Path, PathBuf};
 use std::process::ExitCode;
 use std::time::SystemTime;
@@ -48,30 +47,17 @@ use std::time::SystemTime;
 use anyhow::{Context, Result};
 use cairo_lang_compiler::db::validate_corelib;
 use cairo_lang_compiler::project::{setup_project, update_crate_roots_from_project_config};
-use cairo_lang_defs::db::DefsGroup;
-use cairo_lang_defs::ids::ModuleId;
-use cairo_lang_diagnostics::Diagnostics;
 use cairo_lang_filesystem::db::FilesGroup;
-use cairo_lang_filesystem::ids::{FileId, FileLongId};
-use cairo_lang_lowering::db::LoweringGroup;
-use cairo_lang_lowering::diagnostic::LoweringDiagnostic;
-use cairo_lang_parser::db::ParserGroup;
+use cairo_lang_filesystem::ids::FileLongId;
 use cairo_lang_project::ProjectConfig;
-use cairo_lang_semantic::SemanticDiagnostic;
-use cairo_lang_semantic::db::SemanticGroup;
 use cairo_lang_semantic::plugin::PluginSuite;
-use cairo_lang_utils::Upcast;
-use itertools::Itertools;
 use lsp_server::Message;
-use lsp_types::notification::PublishDiagnostics;
-use lsp_types::{ClientCapabilities, PublishDiagnosticsParams, RegistrationParams, Url};
+use lsp_types::{ClientCapabilities, RegistrationParams};
 use server::connection::ClientSender;
-use state::FileDiagnostics;
-use tracing::{debug, error, info, trace_span, warn};
+use tracing::{debug, error, info, warn};
 
 use crate::config::Config;
 use crate::lang::db::AnalysisDatabase;
-use crate::lang::diagnostics::lsp::map_cairo_diagnostics_to_lsp;
 use crate::lang::lsp::LsProtoGroup;
 use crate::lsp::capabilities::server::{
     collect_dynamic_registrations, collect_server_capabilities,
@@ -299,7 +285,7 @@ impl Backend {
             // Refresh diagnostics each time state changes.
             // Although it is possible to mutate state without affecting the analysis database,
             // we basically never hit such a case in CairoLS in happy paths.
-            scheduler.on_sync_task(Self::refresh_diagnostics);
+            scheduler.on_sync_task(lang::diagnostics::refresh::refresh_diagnostics);
 
             let result = Self::event_loop(&connection, scheduler);
 
@@ -364,203 +350,6 @@ impl Backend {
         }
 
         Ok(())
-    }
-
-    /// Refresh diagnostics and send diffs to the client.
-    #[tracing::instrument(level = "debug", skip_all)]
-    fn refresh_diagnostics(state: &mut State, notifier: Notifier) {
-        // TODO(#6318): implement a pop queue of size 1 for diags
-        let mut files_with_set_diagnostics: HashSet<Url> = HashSet::default();
-        let mut processed_modules: HashSet<ModuleId> = HashSet::default();
-
-        let open_files_ids = trace_span!("get_open_files_ids").in_scope(|| {
-            state
-                .open_files
-                .iter()
-                .filter_map(|uri| state.db.file_for_url(uri))
-                .collect::<HashSet<FileId>>()
-        });
-
-        let open_files_modules =
-            Backend::get_files_modules(&state.db, open_files_ids.iter().copied());
-
-        // Refresh open files modules first for better UX
-        trace_span!("refresh_open_files_modules").in_scope(|| {
-            for (file, file_modules_ids) in open_files_modules {
-                Backend::refresh_file_diagnostics(
-                    state,
-                    &file,
-                    &file_modules_ids,
-                    &mut processed_modules,
-                    &mut files_with_set_diagnostics,
-                    &notifier,
-                );
-            }
-        });
-
-        let rest_of_files = trace_span!("get_rest_of_files").in_scope(|| {
-            let mut rest_of_files: HashSet<FileId> = HashSet::default();
-            for crate_id in state.db.crates() {
-                for module_id in state.db.crate_modules(crate_id).iter() {
-                    if let Ok(module_files) = state.db.module_files(*module_id) {
-                        let unprocessed_files =
-                            module_files.iter().filter(|file| !open_files_ids.contains(file));
-                        rest_of_files.extend(unprocessed_files);
-                    }
-                }
-            }
-            rest_of_files
-        });
-
-        let rest_of_files_modules =
-            Backend::get_files_modules(&state.db, rest_of_files.iter().copied());
-
-        // Refresh rest of files after, since they are not viewed currently
-        trace_span!("refresh_other_files_modules").in_scope(|| {
-            for (file, file_modules_ids) in rest_of_files_modules {
-                Backend::refresh_file_diagnostics(
-                    state,
-                    &file,
-                    &file_modules_ids,
-                    &mut processed_modules,
-                    &mut files_with_set_diagnostics,
-                    &notifier,
-                );
-            }
-        });
-
-        // Clear old diagnostics
-        trace_span!("clear_old_diagnostics").in_scope(|| {
-            let mut removed_files = Vec::new();
-
-            state.file_diagnostics.retain(|uri, _| {
-                let retain = files_with_set_diagnostics.contains(uri);
-                if !retain {
-                    removed_files.push(uri.clone());
-                }
-                retain
-            });
-
-            for file in removed_files {
-                trace_span!("publish_diagnostics").in_scope(|| {
-                    notifier.notify::<PublishDiagnostics>(PublishDiagnosticsParams {
-                        uri: file,
-                        diagnostics: vec![],
-                        version: None,
-                    });
-                });
-            }
-        });
-    }
-
-    /// Refresh diagnostics for a single file.
-    fn refresh_file_diagnostics(
-        state: &mut State,
-        file: &FileId,
-        modules_ids: &Vec<ModuleId>,
-        processed_modules: &mut HashSet<ModuleId>,
-        files_with_set_diagnostics: &mut HashSet<Url>,
-        notifier: &Notifier,
-    ) {
-        let db = &state.db;
-        let file_uri = db.url_for_file(*file);
-        let mut semantic_file_diagnostics: Vec<SemanticDiagnostic> = vec![];
-        let mut lowering_file_diagnostics: Vec<LoweringDiagnostic> = vec![];
-
-        macro_rules! diags {
-            ($db:ident. $query:ident($file_id:expr), $f:expr) => {
-                trace_span!(stringify!($query)).in_scope(|| {
-                    catch_unwind(AssertUnwindSafe(|| $db.$query($file_id)))
-                        .map($f)
-                        .inspect_err(|_| {
-                            error!("caught panic when computing diagnostics for file {file_uri:?}");
-                        })
-                        .unwrap_or_default()
-                })
-            };
-        }
-
-        for module_id in modules_ids {
-            if !processed_modules.contains(module_id) {
-                semantic_file_diagnostics.extend(
-                    diags!(db.module_semantic_diagnostics(*module_id), Result::unwrap_or_default)
-                        .get_all(),
-                );
-                lowering_file_diagnostics.extend(
-                    diags!(db.module_lowering_diagnostics(*module_id), Result::unwrap_or_default)
-                        .get_all(),
-                );
-
-                processed_modules.insert(*module_id);
-            }
-        }
-
-        let parser_file_diagnostics = diags!(db.file_syntax_diagnostics(*file), |r| r);
-
-        let new_file_diagnostics = FileDiagnostics {
-            parser: parser_file_diagnostics,
-            semantic: Diagnostics::from_iter(semantic_file_diagnostics),
-            lowering: Diagnostics::from_iter(lowering_file_diagnostics),
-        };
-
-        if !new_file_diagnostics.is_empty() {
-            files_with_set_diagnostics.insert(file_uri.clone());
-        }
-
-        // Since we are using Arcs, this comparison should be efficient.
-        if let Some(old_file_diagnostics) = state.file_diagnostics.get(&file_uri) {
-            if old_file_diagnostics == &new_file_diagnostics {
-                return;
-            }
-
-            state.file_diagnostics.insert(file_uri.clone(), new_file_diagnostics.clone());
-        };
-
-        let mut diags = Vec::new();
-        let trace_macro_diagnostics = state.config.trace_macro_diagnostics;
-        map_cairo_diagnostics_to_lsp(
-            (*db).upcast(),
-            &mut diags,
-            &new_file_diagnostics.parser,
-            file,
-            trace_macro_diagnostics,
-        );
-        map_cairo_diagnostics_to_lsp(
-            (*db).upcast(),
-            &mut diags,
-            &new_file_diagnostics.semantic,
-            file,
-            trace_macro_diagnostics,
-        );
-        map_cairo_diagnostics_to_lsp(
-            (*db).upcast(),
-            &mut diags,
-            &new_file_diagnostics.lowering,
-            file,
-            trace_macro_diagnostics,
-        );
-
-        trace_span!("publish_diagnostics").in_scope(|| {
-            notifier.notify::<PublishDiagnostics>(PublishDiagnosticsParams {
-                uri: file_uri,
-                diagnostics: diags,
-                version: None,
-            });
-        })
-    }
-
-    /// Gets the mapping of files to their respective modules.
-    fn get_files_modules(
-        db: &AnalysisDatabase,
-        files_ids: impl Iterator<Item = FileId>,
-    ) -> HashMap<FileId, Vec<ModuleId>> {
-        let mut result = HashMap::default();
-        for file_id in files_ids {
-            if let Ok(file_modules) = db.file_modules(file_id) {
-                result.insert(file_id, file_modules.iter().cloned().collect_vec());
-            }
-        }
-        result
     }
 
     /// Calls [`lang::db::AnalysisDatabaseSwapper::maybe_swap`] to do its work.
